3rd Commit
Had to add the displayAvailable() function to "give the user an option
to see a list of how many seats have sold, how many seats are available
in each row, and how many seats are available in the entire auditorium.
" instead of doing that inside the displaySales() function which should
simply output the totalSales variable.